requirements.txt
pandas
lxml
streamlit
google-generativeai
gTTS